ch372: Add default yaserde generator
