+++ src/main.rs
@@ -1,15 +1,22 @@
 mod parser;
 mod generator;
+mod yaserde_generator;
 
 use std::fs;
 use std::io::Read;
 
+
 use crate::parser::parse;
+use crate::yaserde_generator::YaserdeGenerator;
+use crate::generator::Generator;
 
 fn main() {
     let text = load_file("xsd/onvif.xsd");
     if let Ok(f) = parse(text.as_str()) {
-        println!("{}", f)
+        let gen = YaserdeGenerator::new(&f);
+        for ty in f.types {
+            println!("{}", gen.get_rs_entity(&ty));
+        }
     }
 }
